refactor: move derive errors to own mod
