+++ knative-derive/src/inner.rs
@@ -1,5 +1,8 @@
+use crate::{
+    REQUIRED_VARIANTS,
+    error::VerificationError
+};
 use proc_macro::TokenStream;
-use proc_macro2::Span;
 use quote::quote;
 use syn::{
     spanned::Spanned,
@@ -12,35 +15,6 @@ use syn::{
     DeriveInput,
     Ident
 };
-use std::fmt;
-
-const REQUIRED_VARIANTS: [&str; 2] = ["Ready", "Succeeded"];
-
-enum VerificationError {
-    NotDependent(String),
-    OneRequiredVariant,
-}
-
-impl fmt::Display for VerificationError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use VerificationError::*;
-        match self {
-            NotDependent(s) => f.write_fmt(format_args!("{} may not be a dependent", s)),
-            OneRequiredVariant => f.write_fmt(
-                format_args!(
-                    "ConditionType must contain only one of either {} variant",
-                    REQUIRED_VARIANTS.join(" or ")
-                )
-            )
-        }
-    }
-}
-
-impl From<VerificationError> for Error {
-    fn from(v: VerificationError) -> Error {
-        Error::new(Span::call_site(), v)
-    }
-}
 
 fn is_dependent(variant: &syn::Variant) -> bool {
     variant.attrs
